Change model to express effect position from slot id
For certain effects, slot id is not constant, but changes
depending on the effect position. Previously, you cound have
multiple slots with different slot id, but with duplicate
params, except for `fixed_int`, which would get set on
read and be used as a slot match on write.

This worked fine for small amount of param duplication, but
PODxt has quite a lot of params that need to be duplicated,
so this seems like a more efficient method.